Report missing translations when using ActionView::TranslationHelper shortcuts
`I18n.exception_handler` is never invoked when using ActionView::TranslationHelper (ie. `t` in views and helpers). This is because these helpers always set `raise` to true to catch exceptions, which leads I18n to never invoke the exception_handler.

This commit overwrites the built in `t`-method to always re-raise these errors, catches them, and reports them to Rollbar. It then calls translate again with the original options (as to not cause any change in behavior).

NOTE: this assumes that in views and helpers the `t`-shortcut is always used, and never the longer `translate`-method. I checked castupload & prestage, and this is true for both. (Using translate doesn’t cause any problems, it just won’t report to Rollbar for missing translations.)

All of this does not apply to using `I18n.translate` or `I18n.t` directly. Reporting for these methods worked as expected.

Ref https://github.com/denkungsart/prestage/issues/7231